requirements.txt
aiohttp==3.6.3
asn1crypto==1.4.0
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
cycler==0.10.0
discord.py==1.5.1
docutils==0.16
idna==2.10
idna-ssl==1.1.0
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.3
opencv-python==4.4.0.46
Pillow==8.0.1
pycparser==2.20
pygame==2.0.0
Pygments==2.7.2
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
pywin32==300
requests==2.25.0
schedule==0.6.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
wrapt==1.12.1
